chotu_ai/requirements.txt
flask==2.3.3
speechrecognition==3.10.0
pyaudio==0.2.11
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
